app/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
databricks-sdk>=0.99.0
python-multipart==0.0.20
openai==1.82.0
psycopg2-binary==2.9.10
